Run DAP suites with sbt server
