+++ tests/slow/src/test/scala/tests/sbt/SbtStackFrameDapSuite.scala
@@ -1,16 +1,17 @@
-package tests.debug
+package tests.sbt
 
 import scala.meta.internal.metals.debug.Scope
 import scala.meta.internal.metals.debug.Variable
 import scala.meta.internal.metals.debug.Variables
 
-import tests.QuickBuildInitializer
-import tests.QuickBuildLayout
+import tests.SbtBuildLayout
+import tests.SbtServerInitializer
+import tests.debug.BaseStackFrameDapSuite
 
-class StackFrameDapSuite
-    extends BaseStackFrameDapSuite("debug-stack-frame")
-    with QuickBuildInitializer
-    with QuickBuildLayout {
+class SbtStackFrameDapSuite
+    extends BaseStackFrameDapSuite("sbt-debug-stack-frame")
+    with SbtServerInitializer
+    with SbtBuildLayout {
 
   assertStackFrame("foreach")(
     source = """|a/src/main/scala/Main.scala
@@ -23,8 +24,12 @@ class StackFrameDapSuite
                 |  }
                 |}""".stripMargin,
     expectedFrames = List(
-      Variables(Scope.local(Variable("value: int = 1"))),
-      Variables(Scope.local(Variable("value: int = 2")))
+      Variables(
+        Scope.local(Variable("MODULE$: Main$"), Variable("value: int = 1"))
+      ),
+      Variables(
+        Scope.local(Variable("MODULE$: Main$"), Variable("value: int = 2"))
+      )
     )
   )
 
@@ -42,12 +47,19 @@ class StackFrameDapSuite
                 |""".stripMargin,
     expectedFrames = List(
       Variables( // before calculating `z`
-        Scope.local(Variable("x: int = 1"))
+        Scope.local(Variable("MODULE$: Main$"), Variable("x: int = 1"))
       ),
       Variables( // after calculating `z`
-        Scope.local(Variable("x: int = 1"), Variable("z: int = 3"))
+        Scope.local(
+          Variable("MODULE$: Main$"),
+          Variable("x: int = 1"),
+          Variable("z: int = 3")
+        )
       ),
-      Variables(Scope.local(Variable("x$1: Tuple2$mcII$sp")))
+      Variables(
+        Scope.local(Variable("MODULE$: Main$"), Variable("x$1: Tuple2$mcII$sp"))
+      )
     )
   )
+
 }
