+++ tests/slow/src/test/scala/tests/sbt/SbtStepDapSuite.scala
@@ -1,0 +1,10 @@
+package tests.sbt
+
+import tests.SbtBuildLayout
+import tests.SbtServerInitializer
+import tests.debug.BaseStepDapSuite
+
+class SbtStepDapSuite
+    extends BaseStepDapSuite("sbt-debug-step")
+    with SbtServerInitializer
+    with SbtBuildLayout
